Expose similar interface to computeWithDatabase on web - Split up isolate.dart file - Added web extension
